Separated filter tests into a new module
